make getform work (without bcs).
