+++ irksome/backend.py
@@ -1,4 +1,4 @@
-from typing import Protocol
+from typing import Protocol, Any
 import ufl
 from importlib import import_module
 
@@ -6,6 +6,10 @@ from importlib import import_module
 class Backend(Protocol):
     def get_function_space(self, V: ufl.Coefficient) -> ufl.FunctionSpace:
         """Get a function space from the backend"""
+
+    def extract_bcs(bcs: Any)->tuple[Any]:
+        """Extract boundary conditions"""
+
 
     def get_stages(self, V: ufl.FunctionSpace, num_stages: int) -> ufl.Coefficient:
         """
@@ -38,6 +42,14 @@ class Backend(Protocol):
     def get_mesh_constant(MC: MeshConstant | None) -> ufl.core.expr.Expr:
         """Get a backend class to construct a mesh constant from"""
 
+    def TestFunction(space: ufl.FunctionSpace, part: int|None=None)-> ufl.Argument:
+        """Return a test-function that can be used by forms in the backend."""
+
+    def create_nonlinearvariational_problem(F: ufl.Form, g: ufl.Coefficient, solver_parameters: dict):
+        """Create a non-linear variational solver that uses PETSc SNES."""
+
+    def get_stage_spaces(V: ufl.FunctionSpace, num_stages: int) -> ufl.FunctionSpace:
+        """Create a stage space with M number of components."""
 
 def get_backend(backend: str) -> Backend:
     """Get backend class from backend name.
