+++ irksome/backends/dolfinx.py
@@ -2,8 +2,32 @@
 
 try:
     import basix.ufl
-    import dolfinx
+    import dolfinx.fem.petsc
     import ufl
+    import typing
+    import numpy as np
+
+    TestFunction = ufl.TestFunction
+
+    def get_stage_space(V: ufl.FunctionSpace, num_stages:int)->ufl.FunctionSpace:
+        if num_stages == 1:
+            me = V.ufl_elemet()
+        else:
+            el = V.ufl_element()
+            if el.num_sub_elements > 0:
+                me = basix.ufl.mixed_element(np.tile(el.sub_elements, num_stages).tolist())
+            else:
+                me = basix.ufl.blocked_element(el, shape=(num_stages, ))
+        return dolfinx.fem.functionspace(V.mesh, me)
+
+    def extract_bcs(bcs: typing.Any)->tuple[typing.Any]:
+        """Extract boundary conditions"""
+        return bcs
+
+    def create_nonlinearvariational_problem(F: ufl.Form, g: ufl.Coefficient, solver_parameters: dict):
+        """Create a non-linear variational solver that uses PETSc SNES."""
+        return dolfinx.fem.petsc.NonlinearProblem(F, g, petsc_options_prefix="IrkSomeSolver",
+                                                     petsc_options=solver_parameters)
 
     def get_function_space(u: ufl.Coefficient) -> ufl.FunctionSpace:
         return u.ufl_function_space()
